[edit] move to error_gen_function to core_util_box

+++ source/spx_syntax_checker.hpp
@@ -22,12 +22,12 @@ spx_http_syntax_header_line(std::string const& line);
 
 status
 spx_chunked_syntax_start_line(std::string const&				  line,
-							  uint16_t&							  chunk_size,
+							  uint32_t&							  chunk_size,
 							  std::map<std::string, std::string>& chunk_ext);
 
 status
 spx_chunked_syntax_data_line(std::string const&					 line,
-							 uint16_t&							 chunk_size,
+							 uint32_t&							 chunk_size,
 							 std::vector<char>&					 data_store,
 							 std::map<std::string, std::string>& trailer_section);
